+++ fheroes2/src/fheroes2/objects/objtown.cpp
@@ -27,7 +27,7 @@
 
 int ObjTown::GetPassable(u32 index0)
 {
-    u8 index = index0 % 32;
+    u32 index = index0 % 32;
 
     // 13, 29, 45, 61, 77, 93, 109, 125, 141, 157, 173, 189
     if(13 == index || 29 == index)
@@ -42,7 +42,7 @@ int ObjTown::GetPassable(u32 index0)
 
 int ObjTwba::GetPassable(u32 index0)
 {
-    u8 index = index0 % 10;
+    u32 index = index0 % 10;
 
     // 2, 12, 22, 32, 42, 52, 62, 72
     if(2 == index)
